backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic-settings==2.1.0
pydantic==2.5.2
maimai-py==0.1.0
sqlalchemy==2.0.23
redis==5.0.1
aiofiles==23.2.1
python-dotenv==1.0.0
